#121 refactor: api 관련 폴더구조 변경
