+++ src/page/archiving/hook/api/useBlockQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { getDocuments } from '@/shared/api/archiving/document';
+import { getDocuments } from '@/shared/api/time-blocks/team/time-block';
 
 export const useBlockQuery = (teamId: number, blockId: number, selectedId: string) => {
   return useQuery({
