+++ src/page/archiving/hook/api/useTotalDocumentQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { getTotalDocuments } from '@/shared/api/archiving/document';
+import { getTotalDocuments } from '@/shared/api/documents/team/timeline';
 
 export const useTotalDocumentQuery = (teamId: number, type: string, selectedId: string) => {
   return useQuery({
