+++ src/page/archiving/type/documentType.ts
@@ -1,6 +1,6 @@
-export type Document = {
+export interface DocumentType {
   documentId?: number;
   fileName: string;
   fileUrl?: string;
   blockName?: string;
-};
+}
