+++ src/shared/api/documents/team/timeline/index.ts
@@ -1,0 +1,7 @@
+import { axiosInstance } from '@/shared/api/instance';
+
+export const getTotalDocuments = async (teamId: number, type: string) => {
+  const response = await axiosInstance.get(`/documents/team/${teamId}/timeline?type=${type}`);
+
+  return response.data;
+};
