+++ src/shared/api/members/teams/index.ts
@@ -3,7 +3,5 @@ import { axiosInstance } from '@/shared/api/instance';
 export const getClubInfo = async () => {
   const response = await axiosInstance.get(`/members/teams`);
 
-  const clubInfoList = await response.data.data.belongTeamGetResponses;
-
-  return clubInfoList;
+  return response.data;
 };
